\name{pkgIntrogressionv2-package}
\alias{pkgIntrogressionv2-package}
\alias{pkgIntrogressionv2}
\docType{package}
\title{\packageTitle{pkgIntrogressionv2}}
\description{\packageDescription{pkgIntrogressionv2}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{pkgIntrogressionv2}
  \packageIndices{pkgIntrogressionv2}

  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
  \packageAuthor{pkgIntrogressionv2}
  
  Maintainer: \packageMaintainer{pkgIntrogressionv2}
}
\references{
  This optional section can contain literature or other references for
  background information.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
\seealso{
  Optional links to other man pages
}
\examples{
  ## Optional simple examples of the most important functions
  ## Use \dontrun{} around code to be shown but not executed
}
